Update es.json
Hello, I have been a 7th sea Game master for more than 10 years and I have a copy of the manual in Spanish. I propose a series of changes in the terms of some skills and others more in line with the Spanish completion of the game. Delighted of being able to help, thank you for your incredible work with this character sheet.